Lots of changes across the board.
A lot of the build files here slighly modified; when I did the
wyverkiss-static experiment on the master branch (and then made the
entire project its own branch), the master branch was left as a sort of
frankenstein. This commit changes a collection of packages back to
building in a more 'traditional' way, including static libraries but
mostly for foot to use.

No issues were presented when using hikari because it never made any
substantial use of dl_open(). However, wayfire DOES make use of this --
a lot. This is mostly due to its heavily modular nature, accessing
plugins at runtime. A lot of what wayfire does occurs/changes/updates at
runtime, in fact. I think a static wayfire would be... More of an
involved project. One I do not have as much time for.

So, in the meantime, the master branch should build things roughly
canonically, and everything should work ~just fine~. There's a reason
the README says not to use this branch, but I've fixed this lingering
issue :v