chromium: bump to 97.0.4692.71
There's a new segment to the wyverkiss-optional-clang-plugins.patch
which adds --allow-multiple-definition to $LDFLAGS. This is to avoid a
collision in the XDG MIME provided by glib (gio) and the internal XDG
handler in Chromium.

There is potentially a more elegant solution, but the definitions which
clash are exactly identical, so this should be safe.